Fixed recursive relations
When calling recursive relations that resulted with a collection containing other model objects (such as User / Contact) that do not have the same relation name, an exception would be thrown due to the method not existing for the relation call.

This patch properly ensures each resulting model has the relation method to include recursive results.

#108